Refactor signup module structure
  Reorganize signup components and utilities to improve module organization:
  - Move signup blog components to dedicated SignupBlog subdirectory
  - Relocate signup utility to services layer as signup.service.ts
  - Update import paths and create barrel exports for better encapsulation
  - Move and rename test file to match new service structure
