+++ src/app/core/config/views.ts
@@ -2,9 +2,7 @@ import { ComponentClass, FunctionComponent } from 'react';
 
 import { AppView } from '../types';
 
-import AuthView from '../../../views/Signup/components/AuthView';
-import ButtonAuth from '../../../views/Signup/components/ButtonAuth';
-import SignupBlog from '../../../views/Signup/components/SignupBlog';
+import { AuthView, ButtonAuth, SignupBlog } from '../../../views/Signup/components';
 import BlockedAccountView from '../../auth/views/BlockedAccountView/BlockedAccountView';
 import RecoverAccountView from '../../auth/views/RecoverAccountView/RecoverAccountView';
 import RecoveryLinkView from '../../auth/views/RecoveryLinkView/RecoveryLinkView';
