+++ src/views/Signup/components/SignupBlog/index.ts
@@ -1,0 +1,3 @@
+export { default as SignupBlog } from './SignupBlog';
+export { default as AuthView, SignupComponent } from './AuthView';
+export { default as ButtonAuth } from './ButtonAuth';
