+++ src/views/Signup/components/index.ts
@@ -1,8 +1,4 @@
 export { default as SignupForm } from './SignupForm';
-export { default as SignupBlog } from './SignupBlog';
-export { default as AuthView } from './AuthView';
-export { default as ButtonAuth } from './ButtonAuth';
+export { SignupBlog, AuthView, ButtonAuth, SignupComponent } from './SignupBlog';
 export { default as CreateAccountForm } from './CreateAccountForm';
 export { default as WorkspaceGuestSignUp } from './WorkspaceGuestSignUp';
-export { default as ShareGuestSignUpView } from './ShareGuestSignUpView';
-export { SignupComponent } from './AuthView';
