+++ src/views/Signup/services/signup.service.ts
@@ -11,6 +11,7 @@ const postMessage = (data: Record<string, unknown>) => {
   window.top?.postMessage(data, CampaignLinks.PcComponentes);
 };
 
+// TODO: This is duplicating signup logic. Check this in order to refactor and avoid duplication.
 const signup = async (data, dispatch, doRegister, setLoading, appRedirect?, setError?) => {
   if ((data.email === '' && data.password === '') || data.email === null || data.password === null) {
     postMessage({ action: 'autoScroll' });
