+++ src/views/Signup/utils/index.ts
@@ -1,2 +1,1 @@
-export { default as signup } from './signup';
 export { onChangePasswordHandler } from './passwordValidation';
